src/requirements.txt
# --------------------------
# Core Web Framework
# --------------------------
fastapi==0.118.0
uvicorn==0.37.0       # ASGI server
gunicorn==21.2.0      # WSGI server (optional for Render)
slowapi==0.1.9

# --------------------------
# Database
# --------------------------
pymongo==4.15.1       # MongoDB driver
SQLAlchemy==2.0.38    # For relational DB (if used)
psycopg2-binary==2.9.7 # PostgreSQL driver (if using Postgres)

# --------------------------
# Request Parsing & Validation
# --------------------------
python-multipart==0.0.6  # Form data
pydantic==2.11.1         # Data validation
email-validator==2.3.0   # Email field validation

# --------------------------
# Authentication & Security
# --------------------------
PyJWT==2.8.0              # JWT tokens
python-jose==3.3.0        # Optional JWT alternative
passlib[argon2]==1.7.4    # Password hashing

# --------------------------
# Logging
# --------------------------
python-json-logger==4.0.0 # JSON-formatted logs

# --------------------------
# Environment Management
# --------------------------
python-dotenv==1.1.1      # .env support

# --------------------------
# HTTP Requests
# --------------------------
httpx==0.28.1             # Async HTTP requests

# --------------------------
# Paypal
# --------------------------
paypal-server-sdk==1.0.0


# --------------------------
# 2FA / QR Code
# --------------------------
pyotp==2.9.0              # TOTP 2FA codes
qrcode==7.4.2             # QR code generation
Pillow==10.0.0
